renovate: simplify configuration
I believe that `excludePackagePatterns` is the right way to skip packages that
match those patterns during rule evaluation, and still process other gradle
dependencies.

This *should* exclude gradle packages that match (eg.) "^androidx:" etc. but
still process other gradle dependencies that aren't part of the set.

Note: renovate will not do anything with the existing open MRs it created for
the now-excluded packages. We will need to manually close those (and renovate
should not re-open them if this MR is correct)