CI: Configure for scheduled pipeline rebuild of image
The manual mechanism we had configured for building a new docker image meant
that we would have to manually trigger rebuilding this image. This meant that
the image gets out of date quickly from the upstream base-image, introducing new
security issues.

Additionally, the Containerfile digest was not being updated by Renovate, due to
another issue, and so it couldn't even be triggered manually.

This change sets up the CI to run this container rebuild on a schedule instead
of manually. This iss done by adding 'rules' which allow the image build job to
run when its on 'schedule' and the other jobs will be skipped.

The pipeline schedule will be configured in the CI config to run the job on a
weekly basis to keep it up to date.